refactor(formatter/sort-imports): Move `SortImportsOptions` to `sort_imports/options.rs` (#16374)
Pure refactoring.
