+++ crates/oxc_formatter/examples/sort_imports.rs
@@ -3,7 +3,7 @@
 use std::{fs, path::Path};
 
 use oxc_allocator::Allocator;
-use oxc_formatter::{FormatOptions, Formatter, SortImports, SortOrder, get_parse_options};
+use oxc_formatter::{FormatOptions, Formatter, SortImportsOptions, SortOrder, get_parse_options};
 use oxc_parser::Parser;
 use oxc_span::SourceType;
 use pico_args::Arguments;
@@ -21,7 +21,7 @@ fn main() -> Result<(), String> {
     let ignore_case = !args.contains("--no_ignore_case");
     let newlines_between = !args.contains("--no_newlines_between");
 
-    let sort_imports_options = SortImports {
+    let sort_imports_options = SortImportsOptions {
         order,
         partition_by_newline,
         partition_by_comment,
