+++ crates/oxc_formatter/src/ir_transform/mod.rs
@@ -21,4 +21,4 @@
 
 mod sort_imports;
 
-pub use sort_imports::SortImportsTransform;
+pub use sort_imports::*;
