+++ crates/oxc_formatter/src/lib.rs
@@ -16,6 +16,7 @@ use oxc_allocator::Allocator;
 use oxc_ast::ast::*;
 
 pub use crate::embedded_formatter::{EmbeddedFormatter, EmbeddedFormatterCallback};
+pub use crate::ir_transform::options::*;
 pub use crate::options::*;
 pub use crate::service::{oxfmtrc::Oxfmtrc, parse_utils::*};
 use crate::{
